Extracted tracking WooPosItemsSearchAnalyticsTracker
