+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/items/search/WooPosItemsSearchViewModel.kt
@@ -16,15 +16,7 @@ import com.woocommerce.android.ui.woopos.home.items.WooPosItemsViewModel.ItemCli
 import com.woocommerce.android.ui.woopos.home.items.WooPosPaginationState
 import com.woocommerce.android.ui.woopos.home.items.navigation.WooPosItemsNavigator
 import com.woocommerce.android.ui.woopos.home.items.navigation.WooPosItemsNavigator.WooPosItemsScreenNavigationEvent.NavigateToVariationsScreen
-import com.woocommerce.android.ui.woopos.products.GetTotalProductCount
 import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEvent.Event.ItemAddedToCart.WooPosItemSource
-import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEvent.Event.ItemsNextPageLoaded
-import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEvent.Event.PreSearchRecentTermTapped
-import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEvent.Event.SearchRemoteResultsFetched
-import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEventConstant.IS_SEARCH
-import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEventConstant.ITEM_LIST_TYPE
-import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEventConstant.ITEM_LIST_TYPE_PRODUCTS
-import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsTracker
 import com.woocommerce.android.ui.woopos.util.format.WooPosFormatPrice
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Job
@@ -48,8 +40,7 @@ class WooPosItemsSearchViewModel @Inject constructor(
     private val parentToChildrenEventReceiver: WooPosParentToChildrenEventReceiver,
     private val navigator: WooPosItemsNavigator,
     private val searchHelper: WooPosItemsSearchHelper,
-    private val analyticsTracker: WooPosAnalyticsTracker,
-    private val getTotalProductCount: GetTotalProductCount,
+    private val analyticsTracker: WooPosItemsSearchAnalyticsTracker,
 ) : ViewModel() {
     private val _viewState =
         MutableStateFlow<WooPosItemsSearchViewState>(WooPosItemsSearchViewState.Loading)
@@ -87,7 +78,7 @@ class WooPosItemsSearchViewModel @Inject constructor(
                             event.recentSearch
                         )
                     )
-                    trackRecentSearchSelected()
+                    analyticsTracker.trackRecentSearchSelected()
                 }
             }
 
@@ -161,7 +152,7 @@ class WooPosItemsSearchViewModel @Inject constructor(
                     _viewState.value = searchResult.toContentState(searchQuery = query)
                 }
 
-                trackSearchPerformance(searchTimeMillis)
+                analyticsTracker.trackSearchPerformance(searchTimeMillis)
             } else {
                 _viewState.value = WooPosItemsSearchViewState.Error(searchQuery = query)
             }
@@ -213,7 +204,7 @@ class WooPosItemsSearchViewModel @Inject constructor(
         loadMoreJob = viewModelScope.launch {
             val result = dataSource.loadMore(query = currentState.searchQuery)
             _viewState.value = if (result.isSuccess) {
-                trackItemsNextPageLoaded()
+                analyticsTracker.trackItemsNextPageLoaded()
                 result.getOrThrow().toContentState(
                     searchQuery = currentState.searchQuery,
                 )
@@ -221,36 +212,6 @@ class WooPosItemsSearchViewModel @Inject constructor(
                 currentState.copy(paginationState = WooPosPaginationState.Error)
             }
         }
-    }
-
-    private suspend fun trackItemsNextPageLoaded() {
-        val event = ItemsNextPageLoaded.apply {
-            addProperties(
-                mapOf(
-                    ITEM_LIST_TYPE to ITEM_LIST_TYPE_PRODUCTS,
-                    IS_SEARCH to "true"
-                )
-            )
-        }
-        analyticsTracker.track(event)
-    }
-
-    private suspend fun trackRecentSearchSelected() {
-        val event = PreSearchRecentTermTapped.apply {
-            addProperties(
-                mapOf(ITEM_LIST_TYPE to ITEM_LIST_TYPE_PRODUCTS)
-            )
-        }
-        analyticsTracker.track(event)
-    }
-
-    private suspend fun trackSearchPerformance(searchTimeMillis: Long) {
-        val totalProductsCount = getTotalProductCount()
-        val event = SearchRemoteResultsFetched(
-            totalProductsCount = totalProductsCount,
-            millisecondsSinceRequestSent = searchTimeMillis
-        )
-        analyticsTracker.track(event)
     }
 
     private fun handleItemClicked(item: WooPosItemSelectionViewState, source: WooPosItemSource) {
